% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDELIM.R
\name{loadDELIM}
\alias{loadDELIM}
\title{load delimiter-separate files}
\usage{
loadDELIM(
  Dataname = "DELIMdataset",
  Filepath,
  File = NULL,
  Combine = TRUE,
  FileExtension = "csv",
  Header = TRUE,
  Sep = ",",
  Colname = NULL,
  Skip = 0,
  Encoding = ""
)
}
\arguments{
\item{Dataname}{Default "DELIMdataset".
The dataset's name you want.}

\item{Filepath}{The folder path.}

\item{File}{Default \code{NULL}.
Sets one or several files' name with their extension to a vector, or
put all the excel files in the folder into the function in default.}

\item{Combine}{Default \code{TRUE}.
Combines the files to one dataset in default, or loads file each to
be a large list's dataset.}

\item{FileExtension}{Default "csv".
Types in the file extension.}

\item{Header}{Default \code{TRUE}.
Default is setting the first row to column title and read the data from
second row.
Sets FALSE to read the data from the first row and use V1, V2, ... for
column title instead.}

\item{Sep}{Default ",".
Sets the delimiter, and comma in default.}

\item{Colname}{Default \code{NULL}.
Sets a vector for the column title, or gets the original title in default.}

\item{Skip}{Default "0".
Number of rows skipped before reading file, and 0 in default.}

\item{Encoding}{Default "".
Fills in what the file's encoding are, and unknown in default.}
}
\value{
A tibble or tibbles in one list
}
\description{
Loads delimiter-separate files that are in the same folder and with same
format,then combines them into a dataset.
}
\examples{

# There are several files in the test folder called "extdata".
# The folder path should fill in the folder directory of "extdata".
\dontshow{
  path1 <- file.path(system.file("extdata", package = "loadFile"))
  path <- paste0(path1, "/")
  Dataset <- list()
}
# The files include "xlsx", "csv", "txt", and "prn".
# The files with same extension have same format.
# "xlsx" files are excel files.
# "csv" files' data are comma-seperate.
# "txt" files' data are tab-seperate.
# "prn" files' data are fixed-width.

# csv file
# "a.csv" "b.csv" "c.csv"
# Column titles are "Name", "Height", "Weight", and "Score".
# Data starts from the second row.

# Load all csv file that is in the folder and combine them.
loadDELIM(Dataname=c("Test_All1"), Filepath=path)

# Load "a.csv" and "c.csv", then become two tibble in "Test_a&c1".
loadDELIM(Dataname=c("Test_a&c1"), Filepath=path, File=c("a.csv", 
"c.csv"), Combine=FALSE)

# Load all csv file and skip column title, then set the new.
loadDELIM(Dataname=c("Test_Colname1"), Filepath=path, Colname=c("Name_d", 
"H", "W", "score"), Header=FALSE, Skip=1)

# txt file
# "a.txt" "b.txt" "c.txt"
# Variable titles are "Name", "Height", "Weight", and "Score".
# Data starts from the second row.

# Load all txt files that are in the folder and combine them.
loadDELIM(Dataname=c("Test_All2"), Filepath=path, FileExtension="txt", 
Sep="")

# Load "a.txt" and "c.txt", then become two tibble in "Test_data".
loadDELIM(Dataname=c("Test_a&c2"), Filepath=path, File=c("a.txt", "c.txt"), 
Combine=FALSE, FileExtension="txt", Sep="")

# Load all txt files and skip column title, then set the new.
loadDELIM(Dataname=c("Test_Colname2"), Filepath=path, Colname=c("Name_d", 
"H", "W", "score"), FileExtension="txt", Header=FALSE, Sep="", Skip=1)

}
