% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadEXCEL.R
\name{loadEXCEL}
\alias{loadEXCEL}
\title{load excel files}
\usage{
loadEXCEL(
  Dataname = "EXCELdataset",
  Filepath,
  File = FALSE,
  Combine = TRUE,
  Colname = FALSE,
  Skip = 0
)
}
\arguments{
\item{Dataname}{Default "Exceldataset".
The dataset's name you want.}

\item{Filepath}{The folder path.}

\item{File}{Default \code{FALSE}.
Sets one or several files' name with their extension to a vector, or
put all the excel files in the folder into the function in default.}

\item{Combine}{Default \code{TRUE}.
Combines the files to one dataset in default, or loads file each to
be a large list's dataset.}

\item{Colname}{Default \code{FALSE}.
Sets the column title to a vector, or gets the original title in default.}

\item{Skip}{Number of rows skipped before reading file, and 0 in default.}
}
\value{
A tibble or tibbles in one list
}
\description{
Loads excel files that are in the same folder and with same format,
then combines them into a dataset.
}
\examples{

# There are several files in the test folder called "extdata".
# The folder path should fill in the folder directory of "extdata".
\dontshow{
  path1 <- file.path(system.file("extdata", package = "loadFile"))
  path <- paste0(path1, "/")
  Dataset <- list()
}
# The files include "xlsx", "csv", "txt", and "prn".
# The files with same extension have same format.
# "xlsx" files are excel files.
# "csv" files' data are comma-seperate.
# "txt" files' data are tab-seperate.
# "prn" files' data are fixed-width.

# xlsx file
# "a.xlsx" "b.xlsx" "c.xlsx"
# Column titles are "Name", "Height", "Weight", and "Score".
# Data starts from the second row.

# Load all excel files that are in the folder and combine them.
loadEXCEL(Dataname=c("Test_All"), Filepath=path)

# Load "a.xlsx" and "c.xlsx", then become two tibble in "Test_a&c".
loadEXCEL(Dataname=c("Test_a&c"), Filepath=path, File=c("a.xlsx", 
"c.xlsx"), Combine=FALSE)

# Load all excel files and skip column title, then set the new.
loadEXCEL(Dataname=c("Test_Colname"), Filepath=path, Colname=c("Name_d", 
"H", "W", "score"), Skip=1)

}
