% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadWIDTH.R
\name{loadWIDTH}
\alias{loadWIDTH}
\title{load files with fixed-width data}
\usage{
loadWIDTH(
  Dataname = "WIDTHdataset",
  Filepath,
  File = FALSE,
  Combine = TRUE,
  FileExtension = "prn",
  Header = TRUE,
  Width,
  Colname = FALSE,
  Encoding = FALSE
)
}
\arguments{
\item{Dataname}{Default "WIDTHdataset".
The dataset's name you want.}

\item{Filepath}{The folder path.}

\item{File}{Default \code{FALSE}.
Sets one or several files' name with their extension to a vector, or
put all the excel files in the folder into the function in default.}

\item{Combine}{Default \code{TRUE}.
Combines the files to one dataset in default, or loads file each to
be a large list's dataset.}

\item{FileExtension}{Default "prn".
Types in the file extension.}

\item{Header}{Default \code{TRUE}.
Default is setting the first row to column title and read the data from
second row.
Sets FALSE to read the data from the first row and use V1, V2,... for
column title instead.}

\item{Width}{Sets the width for each variable to a vector.}

\item{Colname}{Default \code{FALSE}.
Sets the column title to a vector, or gets the original title in default.}

\item{Encoding}{Default \code{FALSE}.
Fills in what the file's encoding are, and unknown in default.}
}
\value{
A tibble or tibbles in one list
}
\description{
Loads files that with fixed-width data, in the same folder and with same
format,then combines them into a dataset.
}
\examples{

# There are several files in the test folder called "extdata".
# The folder path should fill in the folder directory of "extdata".
\dontshow{
  path1 <- file.path(system.file("extdata", package = "loadFile"))
  path <- paste0(path1, "/")
  Dataset <- list()
}
# The files include "xlsx", "csv", "txt", and "prn".
# The files with same extension have same format.
# "xlsx" files are excel files.
# "csv" files' data are comma-seperate.
# "txt" files' data are tab-seperate.
# "prn" files' data are fixed-width.

# prn file
# "a.prn" "b.prn" "c.prn"
# The data are "Name", "Height", "Weight", and "Score".
# Data widths are 6, 3, 2, 1 in order.
# There are no title row in the data. 
# Data starts from the first row.

# Load all prn file that is in the folder and combine them.
loadWIDTH(Dataname=c("Test_All"), Filepath=path, Header=FALSE, 
Width=c(6, 3, 2, 1))

# Load "a.prn" and "c.prn", then become two tibble in "Test_a&c1".
loadWIDTH(Dataname=c("Test_a&c"), Filepath=path, Header=FALSE, 
File=c("a.csv", "c.csv"), Combine=FALSE, Width=c(6, 3, 2, 1))

# Load all csv file and skip column title, then set the new.
loadWIDTH(Dataname=c("Test_Colname"), Filepath=path, Header=FALSE, 
Width=c(6, 3, 2, 1), Colname=c("Name_d", "H", "W", "score"))

}
